+++ invenio_communities/communities/resources/config.py
@@ -24,7 +24,7 @@ from invenio_communities.communities.resources.serializer import (
 from invenio_communities.errors import (
     CommunityFeaturedEntryDoesNotExistError,
     LogoSizeLimitError,
-    OpenRequestsForCommunityError,
+    OpenRequestsForCommunityDeletionError,
 )
 
 community_error_handlers = RecordResourceConfig.error_handlers.copy()
@@ -48,7 +48,7 @@ community_error_handlers.update(
                 description=str(e),
             )
         ),
-        OpenRequestsForCommunityError: create_error_handler(
+        OpenRequestsForCommunityDeletionError: create_error_handler(
             lambda e: HTTPJSONException(
                 code=400,
                 description=str(e),